feat(cabi): Group D.1 — async-reject + CKA_ID find (177→182)
GAsync (1): C_OpenSession now rejects CKF_ASYNC_SESSION with
CKR_SESSION_ASYNC_NOT_SUPPORTED (0x205) per §5.6.1; the token never advertises
async support. Adds the two missing constants.

CkaIdRetrieval (4): store_ulong now writes CK_ULONG attributes at native width
(size_of::<usize>(): 4 bytes wasm32, 8 bytes native) instead of a fixed 4.
CKA_CLASS is server-managed so absorb_template_attrs skips it — only store_ulong
set it, at 4 bytes — while a native C_FindObjects filter supplies CKA_CLASS as
an 8-byte CK_ULONG, so the byte-exact match missed every generated key. Native
width makes generated objects compare equal to caller templates and read back at
native width via C_GetAttributeValue. All map readers take the low 4 bytes
(from_le_bytes([v[0..3]])), so the widening is backward-compatible; the only
u64 reader (CKA_LEAF_INDEX) does not overlap store_ulong attrs.

Gate: compliance 182 PASS / 45 FAIL / 26 SKIP (no crash); interop 8+2+5 = 15/15.

+++ rust/src/state.rs
@@ -788,9 +788,15 @@ pub fn store_bool(attrs: &mut Attributes, attr_type: u32, value: bool) {
     attrs.insert(attr_type, vec![if value { 0x01 } else { 0x00 }]);
 }
 
-/// Store a CK_ULONG attribute (4-byte little-endian).
+/// Store a CK_ULONG attribute at native CK_ULONG width (4 bytes on wasm32,
+/// 8 bytes on 64-bit native — `size_of::<usize>()`, matching the C-ABI
+/// marshaling). This keeps generated objects byte-compatible with caller
+/// templates: a `CKA_CLASS`/`CKA_KEY_TYPE` find-filter supplied as a native
+/// CK_ULONG compares byte-exact in `C_FindObjects`, and the value reads back at
+/// native width through `C_GetAttributeValue`. All map readers take the low
+/// 4 bytes (`from_le_bytes([v[0..3]])`), so widening is backward-compatible.
 pub fn store_ulong(attrs: &mut Attributes, attr_type: u32, value: u32) {
-    attrs.insert(attr_type, value.to_le_bytes().to_vec());
+    attrs.insert(attr_type, (value as usize).to_le_bytes().to_vec());
 }
 
 /// Read a CK_BBOOL attribute back from an attrs HashMap (returns false if absent).
